HW1/hw1/src/main/java/tqs/homework/hw1/services/ReservationService.java
package tqs.homework.hw1.services;

import java.util.List;

import org.springframework.stereotype.Service;

import tqs.homework.hw1.models.Reservation;
import tqs.homework.hw1.repositories.ReservationRepository;

@Service
public class ReservationService {

    private final ReservationRepository reservationRepository;

    public ReservationService(ReservationRepository reservationRepository) {
        this.reservationRepository = reservationRepository;
    }

    public Reservation getReservationById(Long id) {
        return reservationRepository.findById(id).orElse(null);
    }

    public Reservation saveReservation(Reservation reservation) {
        return reservationRepository.save(reservation);
    }

    public void deleteReservation(Reservation reservation) {
        reservationRepository.delete(reservation);
    }

    public List<Reservation> getAllReservations() {
        return reservationRepository.findAll();
    }
    
}
